core: fix REQUEST_MESSAGE drops with multiple components
We often flagged REQUEST_MESSAGE commands as Busy because they were
duplicate. However, when we are targetting different component IDs, they
are obviously different and need to be done for this to work properly.

This came up when debugging a gimbal not being recognized as the request
for GIMBAL_DEVICE_INFORMATION was swallowed.