mavlink_direct: fix message filtering
The message filtering was not correct with 2 or more subscriptions as
they would conflict with each other. This should now be fixed with each
subscription keeping track of what to filter out.

While doing so we switched to the new and nicer Handle/Subscription API
in SystemImpl rather than the ugly void* cookie thing.