core/plugins: make unregister blocking in dtor
This avoid potential race conditions on plugin destruction. Basically,
this avoids the case where we want to unregister but it doesn't fully
unregister but only adds it to the list to "unregister later" while
there is already a mavlink message being processed.

By using the blocking method, we don't use the "unregister later" list,
which should be safe for plugin destruction.

The fix is applied to all plugins. However, the problem occurred only
with the events plugin in the PX4 MAVSDK integration tests.